% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bonnie.f.R
\name{bonnie.f}
\alias{bonnie.f}
\title{\code{bonnie.f} Calculate nested F-ratios - "correctly" (for Bonnie
Waring, anyway)}
\usage{
bonnie.f(model, high, low)
}
\arguments{
\item{model}{standard linear model to be calculated}

\item{high}{Highest hierarchical-level in model}

\item{low}{Lowest hierarchical-level in model}
}
\value{
Numeric
}
\description{
\code{bonnie.f} Calculate nested F-ratios - "correctly" (for Bonnie
Waring, anyway)
}
\details{
Divides the F-value of the highest level by that of the interaction between it and the lowest-value. No p-value, you can do that yourself!
}
\author{
Will Pearse
}

