% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phylo.manip.R
\name{drip.node.labels}
\alias{drip.node.labels}
\title{\code{drip.node.labels} 'Drip' node.labels through a phylogeny}
\usage{
drip.node.labels(tree, clade.mat = NULL)
}
\arguments{
\item{tree}{ape::phylo object; must have node.label slot (otherwise there're no node labels to use!)}

\item{clade.mat}{If not NULL (the default), the \code{clade.mat} slot of a caper::clade.matrix function call on the phylogeny to be labelled. You may already have this lying around, if so it speeds the function somewhat.}
}
\value{
Vector of labels; could be put straight into an ape::phylo object in the node.label slot.
}
\description{
\code{drip.node.labels} 'Drip' node.labels through a phylogeny
}
\details{
Returns a vector that names all nodes according to the label of their ancestral nodes. So if the Fagaceae node is labelled, all nodes within that clade will be labelled Fagaceae also. A new node label, such a Quercus, will be used instead if one is encountered.
}
\note{
Convenient if you do a lot of OU modelling...
}
\author{
Will Pearse
}

