\name{find.unique.branch.length}
\alias{find.unique.branch.length}
\title{\code{find.unique.branch.length} Find the terminal edge length leading to a species in a phylogeny}
\usage{
find.unique.branch.length(tree, tip)
}
\arguments{
  \item{tree}{the phylogeny to be searched}

  \item{tip}{the tip whose edge length must be found}
}
\value{
The edge length leading to that tip
}
\description{
\code{find.unique.branch.length} Find the terminal edge
length leading to a species in a phylogeny
}
\details{
If you find yourself manually manipulating phylogenies a
lot, this is for you
}
\author{
Will Pearse
}

