% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phylo.manip.R
\name{label.tree.nodes}
\alias{label.tree.nodes}
\title{\code{label.tree.nodes} Label nodes of phylogeny according to tip labels}
\usage{
label.tree.nodes(tree, groups, clade.mat = NULL)
}
\arguments{
\item{tree}{ape::phylo object; must have node.label slot (otherwise there're no node labels to use!)}

\item{groups}{The groups to be painted onto the nodes}

\item{clade.mat}{If not NULL (the default), the \code{clade.mat} slot of a caper::clade.matrix function call on the phylogeny to be labelled. You may already have this lying around, if so it speeds the function somewhat.}
}
\value{
Vector of labels; could be put straight into an ape::phylo object in the node.label slot.
}
\description{
\code{label.tree.nodes} Label nodes of phylogeny according to tip labels
}
\details{
Returns a named vector describing what the node.label of a phylogeny should be to match the tips. It's essentially a ppor man's ancestral state reconstruction, such that you can paint values for species onto the clade that contains them all on a phylogeny. For instance, if you know thirty species are in Fagaceae, it will paint those labels onto a clade subtending from them.
}
\note{
A friend of \code{label.tree.tips}
}
\author{
Will Pearse
}

