% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phylo.manip.R
\name{label.tree.tips}
\alias{label.tree.tips}
\title{\code{label.tree.tips} Label tips of phylogeny according to node labels}
\usage{
label.tree.tips(tree, clade.mat = NULL)
}
\arguments{
\item{tree}{ape::phylo object; must have node.label slot (otherwise there're no node labels to use!)}

\item{clade.mat}{If not NULL (the default), the \code{clade.mat} slot of a caper::clade.matrix function call on the phylogeny to be labelled. You may already have this lying around, if so it speeds the function somewhat.}
}
\value{
Named vector describing tree.
}
\description{
\code{label.tree.tips} Label tips of phylogeny according to node labels
}
\details{
Returns a named vector describing what node.label the tips in your phylogeny subtends from. Each tip will only be labelled by one node - the youngest. This is essentially a way of figuring out if a species is in a particular clade (e.g., Fagaceae) from \code{node.label}s in a phylogeny.
}
\author{
Will Pearse
}

