% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lean}
\alias{lean}
\title{\code{lean} Local Environmental Adaptation Null model}
\usage{
lean(union = FALSE)
}
\arguments{
\item{union}{Allow for juvenile adaptation}
}
\value{
Nothing; plot-type
}
\description{
\code{lean} Local Environmental Adaptation Null model
}
\details{
This is extremely unlikely to be of any interest to anyone
other than those attending the Silwood Park Christmas 2010 meeting,
where the importance of LEAN was disucssed in relation to student
teacher. Seriously, don't even bother running it.
}
\examples{
\dontrun{
lean()
lean(FALSE)
}
}
\author{
Will Pearse
}

