% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.transition.calc}
\alias{plot.transition.calc}
\title{\code{transition.calc} regression of aged transitions among character states}
\usage{
\method{plot}{transition.calc}(x, ...)
}
\description{
\code{transition.calc} regression of aged transitions among character states
}

