% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gis.R
\name{read.ncdc}
\alias{read.ncdc}
\title{Load US NCDC data (climate, etc.)}
\usage{
read.ncdc(file, tempout = "delete.me.txt")
}
\arguments{
\item{file}{location (on hard drive or URL) of NCDC file to be loaded}

\item{tempout}{name of temporary file to be written out (and then
deleted) as part of read process. If this file already exists,
it'll get overwritten and then deleted!}
}
\description{
Load US NCDC data (climate, etc.)
}
\examples{
palmer <- read.ncdc("http://www1.ncdc.noaa.gov/pub/data/cirs/climdiv/climdiv-pdsidv-v1.0.0-20160105")
}

