% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trait.space}
\alias{trait.space}
\title{\code{trait.space} (maybe) make a trait space of species co-existence}
\usage{
trait.space(comm, abundance = FALSE, plot = TRUE, coexist = NULL,
  mask = NULL, ...)
}
\arguments{
\item{comm}{Community matrix (communities in rows, species in column); NAs would probably be a bad idea}

\item{abundance}{Work with species abundances? Default FALSE}

\item{plot}{Produce a plot of the space at the end of the run? Default TRUE}

\item{coexist}{Provide a co-existence matrix to use for fitting (see below). Default to NULL to calculate one as part of the run. The matrix will be returned at the end of the run so you can work with it later. It is *definitely* worth calculating this once, as I've almost maliciously not optimised this code.}

\item{mask}{matrix of TRUE and FALSE, where only elements of the coexistence matrix where this is TRUE will be used in the calculations.}

\item{...}{additional arguments to be passed to the optimisation routine. See the notes for advice on this.}
}
\value{
The data that were plotted last, the PCA and clade.matrix to speed later plots, and the colours used.
}
\description{
\code{trait.space} (maybe) make a trait space of species co-existence
}
\details{
Attempts to place all species in a 2-D space where distance between species is directly related to their proportional coexistence. I.e., if species are >=1 unit distant from each other, they never coexit, if they're on top of one-another, they always co-exist. See notes.
}
\note{
It is *strongly* advised that you check the convergence of this routine, and play around with the algorithm used to assess convergence. This came about as a result of a NutNet (http://www.nutnet.umn.edu/) meeting; please don't run away and use this without bringing me on-board, or at the very least informing the organisers of NutNet. Please don't become "that guy" and ignore this...
}
\author{
Will Pearse
}

