% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transition.regression.R
\name{transition.calc}
\alias{transition.calc}
\title{\code{transition.calc} regression of aged transitions among character states}
\usage{
transition.calc(tree, continuous, discrete, simmap.model = "ER",
  simmap.nsim = 1000, simmap.pi = "estimated")
}
\arguments{
\item{tree}{a phylogeny (of class phylo)}

\item{continuous}{continuous trait (named) to be regressed}

\item{discrete}{discrete category (named) species are evolving
between whose transitions you wish to model}

\item{simmap.model}{parameter for \code{\link[phytools]{make.simmap}}}

\item{simmap.nsim}{parameter for \code{\link[phytools]{make.simmap}}}

\item{simmap.pi}{parameter for \code{\link[phytools]{make.simmap}}}
}
\description{
\code{transition.calc} regression of aged transitions among character states
}
\examples{
\dontrun{
tree <- rtree(50)
continuous <- rnorm(50); names(continuous) <- tree$tip.label
discrete <- factor(sample(letters[1:6], 50, replace=TRUE)); names(discrete) <- tree$tip.label
silly <- transition.calc(tree, continuous, discrete, simmap.nsim=10)
plot(silly)
}
}
\author{
Will Pearse
}

