\name{willeerd-package}
\alias{willeerd-package}
\alias{willeerd}
\docType{package}
\title{
Things that Will has written in R
~~ willeerd ~~
}
\description{
I work on ecology/evolutionary biology, and the entries in this package likely reflect that.
The code in this is far from production-ready; you would be lucky if it were tested.
I'm releasing this online in the spirit of being helpful to other people. Please, don't be that person who takes/publishes from this without at the very least asking me to help out first.
}
\details{
\tabular{ll}{
Package: \tab willeerd\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-11-22\cr
License: \tab GPL with the following addendum: use of this code in a publication and/or use of derivatives or inspiration taken from this code without the explicit consent of me (Will Pearse) are not allowed. Talk to me: I'm a nice guy, and you'll be better for it, honest. Also, there are bugs-a-plenty in here!
}
}
\author{
Will Pearse

Maintainer: Will Pearse! If you want to complain, go fork me :p
}
\references{
Ha!
}
\keyword{ package }
\examples{
#Ha!
}
