% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phylo.plots.R
\name{willeerd.nodelabels}
\alias{willeerd.nodelabels}
\title{\code{willeerd.nodelabels} Plot tip labels with radial spacing}
\usage{
willeerd.nodelabels(text, node, radial.adj = 1, adj = c(0.5, 0.5),
  frame = "rect", pch = NULL, thermo = NULL, pie = NULL,
  piecol = NULL, col = "black", bg = "lightblue", horiz = FALSE,
  width = NULL, height = NULL, ...)
}
\arguments{
\item{text}{see \code{\link[ape]{nodelabels}}}

\item{node}{which node(s) to plot}

\item{radial.adj}{a multiplier for how far out each node label
should be}

\item{adj}{see \code{\link[ape]{nodelabels}}}

\item{frame}{see \code{\link[ape]{nodelabels}}}

\item{pch}{see \code{\link[ape]{nodelabels}}}

\item{thermo}{see \code{\link[ape]{nodelabels}}}

\item{pie}{see \code{\link[ape]{nodelabels}}}

\item{piecol}{see \code{\link[ape]{nodelabels}}}

\item{col}{see \code{\link[ape]{nodelabels}}}

\item{bg}{see \code{\link[ape]{nodelabels}}}

\item{horiz}{see \code{\link[ape]{nodelabels}}}

\item{width}{see \code{\link[ape]{nodelabels}}}

\item{height}{see \code{\link[ape]{nodelabels}}}

\item{...}{see \code{\link[ape]{nodelabels}}}
}
\value{
...exactly as nodelabels
}
\description{
\code{willeerd.nodelabels} Plot tip labels with radial spacing
}
\details{
A way of getting evenly spaced node labels on a radial
phylogeny. Just run the examples, it's quite simple. This is a very
minorly-editted version of nodelabels; I can't take much credit!
Don't cite this, cite ape!
}
\examples{
\dontrun{
tree <- stree(128, type="balanced")
plot(tree, type="radial", show.tip.label=FALSE)
willeerd.nodelabels(pch=20)
willeerd.nodelabels(pch=20, adj=1.05, col="red")
willeerd.nodelabels(pch=20, adj=1.1, col="blue")
}
}
\author{
Will Pearse
}

