% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transition.regression.R
\name{plot.transition.calc}
\alias{plot.transition.calc}
\title{\code{transition.calc} regression of aged transitions among character states}
\usage{
\method{plot}{transition.calc}(x, ...)
}
\description{
\code{transition.calc} regression of aged transitions among character states
}

